requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
greenlet==3.0.1
idna==3.6
iniconfig==2.0.0
packaging==23.2
playwright==1.40.0
pluggy==1.3.0
pyee==11.0.1
pytest==7.4.4
pytest-base-url==2.0.0
pytest-playwright==0.4.3
python-slugify==8.0.1
requests==2.31.0
text-unidecode==1.3
typing_extensions==4.9.0
urllib3==2.1.0
